ref: API へのアクセス処理を  src/script/ へ外部化
