+++ src/pages/api/movies.ts
@@ -1,18 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
-import type { Movie, MovieId, MovieRevenue } from "types/movie";
-import movieIdRevenueMaps from 'consts/movie-Id-revenue-map'
+import type { Movie } from "types/movie";
 
-const movieIdList = (index: number): MovieId[] => {
-  return Array.from(movieIdRevenueMaps[index].keys())
-}
-
-const getMovieData = (index: number): Promise<Movie>[] => {
-  return movieIdList(index).map((id) => {
-    return fetch(
-      `https://api.themoviedb.org/3/movie/${id}?api_key=${process.env.TMDB_API_KEY}&language=ja`
-    ).then(res => res.json())
-  })
-}
+import { getMovieData } from "script/get-movie"
 
 const handler = async (req: NextApiRequest, res: NextApiResponse<Movie[]>) => {
   const indexNumber = Number(req.query.page) - 1
